add generic DOI
not sure what the best practice is, but I can't get the new doi until I make a new release so if i don't change it it'll have the old doi stuck in the release version. this way it'll have the doi that points to the latest in the release archive and i can update the master (which is what people tend to download) with the specific doi after the release? is there a better way to do this?